requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile ./requirements.in
#
alembic==1.13.1
    # via
    #   -r ./requirements.in
    #   flask-migrate
aniso8601==9.0.1
    # via flask-restful
apispec==6.3.1
    # via
    #   -r ./requirements.in
    #   flask-apispec
apscheduler==3.10.4
    # via
    #   -r ./requirements.in
    #   flask-apscheduler
astroid==3.0.2
    # via pylint
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
blinker==1.7.0
    # via flask
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   -r ./requirements.in
    #   flask
    #   safety
coverage[toml]==7.4.0
    # via
    #   -r ./requirements.in
    #   pytest-cov
dill==0.3.7
    # via pylint
dparse==0.6.3
    # via safety
et-xmlfile==1.1.0
    # via openpyxl
flake8==6.1.0
    # via -r ./requirements.in
flask==3.0.0
    # via
    #   -r ./requirements.in
    #   flask-apispec
    #   flask-apscheduler
    #   flask-cors
    #   flask-jwt-extended
    #   flask-marshmallow
    #   flask-migrate
    #   flask-restful
    #   flask-sqlalchemy
flask-apispec==0.11.4
    # via -r ./requirements.in
flask-apscheduler==1.13.1
    # via -r ./requirements.in
flask-cors==4.0.0
    # via -r ./requirements.in
flask-jwt-extended==4.6.0
    # via -r ./requirements.in
flask-marshmallow==0.15.0
    # via -r ./requirements.in
flask-migrate==4.0.5
    # via -r ./requirements.in
flask-restful==0.3.10
    # via -r ./requirements.in
flask-sqlalchemy==3.1.1
    # via
    #   -r ./requirements.in
    #   flask-migrate
greenlet==3.0.3
    # via sqlalchemy
idna==3.6
    # via requests
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via
    #   -r ./requirements.in
    #   pylint
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via
    #   -r ./requirements.in
    #   flask
jsonschema==4.20.0
    # via -r ./requirements.in
jsonschema-specifications==2023.12.1
    # via
    #   -r ./requirements.in
    #   jsonschema
mako==1.3.0
    # via alembic
markupsafe==2.1.3
    # via
    #   -r ./requirements.in
    #   jinja2
    #   mako
    #   werkzeug
marshmallow==3.20.1
    # via
    #   -r ./requirements.in
    #   flask-apispec
    #   flask-marshmallow
    #   marshmallow-enum
    #   marshmallow-sqlalchemy
    #   webargs
marshmallow-enum==1.5.1
    # via -r ./requirements.in
marshmallow-sqlalchemy==0.29.0
    # via -r ./requirements.in
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
openpyxl==3.1.2
    # via -r ./requirements.in
packaging==21.3
    # via
    #   apispec
    #   dparse
    #   flask-marshmallow
    #   marshmallow
    #   marshmallow-sqlalchemy
    #   pytest
    #   safety
    #   webargs
platformdirs==4.1.0
    # via pylint
pluggy==1.3.0
    # via pytest
psycopg2==2.9.9
    # via -r ./requirements.in
pycodestyle==2.11.1
    # via flake8
pyflakes==3.1.0
    # via flake8
pyjwt==2.8.0
    # via
    #   -r ./requirements.in
    #   flask-jwt-extended
pylint==3.0.3
    # via -r ./requirements.in
pyparsing==3.1.1
    # via packaging
pytest==7.4.4
    # via
    #   -r ./requirements.in
    #   pytest-cov
    #   pytest-mock
pytest-cov==4.1.0
    # via -r ./requirements.in
pytest-mock==3.12.0
    # via -r ./requirements.in
python-dateutil==2.8.2
    # via
    #   -r ./requirements.in
    #   flask-apscheduler
python-dotenv==1.0.0
    # via -r ./requirements.in
pytz==2023.3.post1
    # via
    #   apscheduler
    #   flask-restful
pyyaml==6.0.1
    # via responses
referencing==0.32.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   -r ./requirements.in
    #   responses
    #   safety
responses==0.24.1
    # via -r ./requirements.in
rpds-py==0.16.2
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.5
    # via safety
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
safety==2.3.5
    # via -r ./requirements.in
six==1.16.0
    # via
    #   apscheduler
    #   flask-restful
    #   python-dateutil
sqlalchemy==2.0.25
    # via
    #   -r ./requirements.in
    #   alembic
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
tomlkit==0.12.3
    # via pylint
typing-extensions==4.9.0
    # via
    #   alembic
    #   sqlalchemy
tzlocal==5.2
    # via apscheduler
urllib3==2.1.0
    # via
    #   -r ./requirements.in
    #   requests
    #   responses
webargs==8.3.0
    # via flask-apispec
werkzeug==3.0.1
    # via
    #   -r ./requirements.in
    #   flask
    #   flask-jwt-extended
wheel==0.42.0
    # via -r ./requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
